refactor: camera -> camera location
