+++ src/pages/Project/EditProject/CameraLocation.vue
